Time: 0 ms (100.00%) | Memory: 8.6 MB (70.86%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -7,6 +7,6 @@ public:
         return x*solve(x*x,(n-1)/2);
     }
     double myPow(double x, int n) {
-        return solve(x,(long)n);
+        return solve(x,n);
     }
 };
